requirements.txt
gspread
oauth2client
python-jobspy
pandas
google-genai
pyyaml
python-dotenv
openai
requests
beautifulsoup4
pytest
